+++ TextSecureiOS/ViewControllers/VerificationCodeViewController.m
@@ -101,7 +101,7 @@
                 
             default:
                 [[[UIAlertView alloc]initWithTitle:@"Can't verify" message:@"An unknown error occured. Pleasy try again." delegate:self cancelButtonTitle:@"Ok" otherButtonTitles:nil, nil] show];
-                DLog(@"Verification operation failed with response: %@ and response code : %li", responseObject, operation.response.statusCode);
+                DLog(@"Verification operation failed with response: %@ and response code : %li", responseObject, (long)operation.response.statusCode);
                 break;
         }
